Added frontend components for better UI and tested the payEmployees and purchase Vehicle flows

+++ backend/DjangoServer/autoshop/models.py
@@ -30,7 +30,7 @@ class AutoUser(models.Model):
     needHelp = models.BooleanField(default=False)
     ethicsViolation = models.TextField(default="None")
     location = models.CharField(max_length=50, default="unknown")
-    hoursOwed = models.IntegerField(default=0)
+    hoursOwed = models.FloatField(default=0)
     wage = models.FloatField(default=7.25)
 
 class Reservation(models.Model):
